fixed testRunnerReader lint error
