+++ kafka/consumer/consumer_auto_test.go
@@ -2,9 +2,11 @@ package consumer
 
 import (
 	"context"
+	"math/rand"
 	"testing"
 	"time"
 
+	"github.com/google/uuid"
 	"github.com/segmentio/kafka-go"
 	"github.com/stretchr/testify/assert"
 )
@@ -61,4 +63,42 @@ func (c *packageAutoMock) Consume(ctx context.Context, topic string) (<-chan Mes
 	ac := newAutoConsumer(topic, c.brokers, c.opts...)
 	go ac.run(ctx)
 	return ac.channel, func() error { return ac.stop() }
+}
+
+type testRunnerReader struct {
+	topic string
+}
+
+func newTestRunnerReader(topic string) *testRunnerReader {
+	return &testRunnerReader{
+		topic: topic,
+	}
+}
+
+func (trr *testRunnerReader) ReadMessage(ctx context.Context) (kafka.Message, error) {
+	msg := trr.newMessage()
+	return msg, nil
+}
+
+func (trr *testRunnerReader) FetchMessage(ctx context.Context) (kafka.Message, error) {
+	msg := trr.newMessage()
+	return msg, nil
+}
+
+func (trr *testRunnerReader) CommitMessages(ctx context.Context, msgs ...kafka.Message) error {
+	return nil
+}
+
+func (trr *testRunnerReader) Close() error {
+	return nil
+}
+
+func (trr *testRunnerReader) newMessage() kafka.Message {
+	msg := kafka.Message{}
+
+	msg.Topic = trr.topic
+	msg.Offset = rand.Int63()           //#nosec G404 -- This for the test runner reader, not used in production.
+	msg.Partition = rand.Intn(20-1) + 1 //#nosec G404 -- This for the test runner reader, not used in production.
+	msg.Value = []byte(uuid.New().String())
+	return msg
 }
